[CAP-004] Refactor,  SummaryRequestVC구현
여러타입에 익스텐셥앱에서 해당 뷰컨트롤러를 사용하기위해 구현하였습니다.
